requirements.txt
# Core
pydantic>=2.10,<3.0
PyYAML>=6.0,<7.0

# LLM clients
openai>=1.50,<2.0
langchain-openai>=0.3,<1.0
langchain-core>=0.3,<1.0
langgraph>=0.3,<1.0

# Vector store
chromadb>=0.6,<1.0

# Document processing
PyMuPDF>=1.25,<2.0
python-docx>=1.1,<2.0
Pillow>=11.0,<12.0

# Knowledge graph
networkx>=3.4,<4.0

# Web UI
streamlit>=1.41,<2.0
streamlit-agraph>=0.0.45

# Utilities
posthog<4
